+++ lib/index.ts
@@ -1,3 +1,10 @@
+/* tslint:disable:import-spacing */
 import VueSlider from './vue-slider'
-export { ERROR_TYPE } from './utils/control'
+import VueSliderMark from './vue-slider-mark'
+import VueSliderDot from './vue-slider-dot'
+import { ERROR_TYPE } from './utils/control'
+;(VueSlider as any).VueSliderMark = VueSliderMark
+;(VueSlider as any).VueSliderDot = VueSliderDot
+
 export default VueSlider
+export { ERROR_TYPE, VueSliderMark, VueSliderDot }